+++ darts/models/forecasting/xgboost.py
@@ -368,3 +368,8 @@ class XGBModel(RegressionModel, _LikelihoodMixin):
                 else self.output_chunk_length
             ),
         )
+
+    @property
+    def _supports_native_multioutput(self):
+        # since xgboost==2.1.0, likelihoods do not support native multi output regression
+        return super()._supports_native_multioutput and self.likelihood is None